Mechanim module now adds Animator.StringToHash ints
It will automatically add them to all lines that the Mecanim module picks up, with the same name plus "Hash" appended. Added methods to get the netstandard and unityengine dlls in order for the compiler to complete successfully. This version includes the animation controller layer names and sub state names (fullPathHash). The string searching logic could be improved here.